requirements.txt
python-jobspy
pandas
pyyaml
apscheduler
openpyxl
customtkinter
ddgs
